Change name PicturePartManager into PostPictureManager, merge class PartManager inot PostManager, merge class PartController into PostController
