+++ src/View/Part/_form.html.twig
@@ -26,7 +26,7 @@
                         <label for="category"></label>
                         <select name="category" id="category_select" required>
                             <option value="">CATEGORY</option>
-                        {% for key_category, category in constant('App\\Model\\PartManager::CATEGORY') %}
+                        {% for key_category, category in constant('App\\Model\\PostManager::CATEGORY') %}
                             <option value="{{ key_category + 1 }}">{{ category }}</option>
                         {% endfor %}
                         </select>    
@@ -35,7 +35,7 @@
                         <label for="wear"></label>
                         <select name="wear" id="wear_select" required>
                             <option value="">WEAR STATE</option>
-                        {% for key, wear in constant('App\\Model\\PartManager::WEAR') %}
+                        {% for key, wear in constant('App\\Model\\PostManager::WEAR') %}
                             <option value="{{ key }}">{{ wear }}</option>
                         {% endfor %}
                         
@@ -47,7 +47,7 @@
                         <label for="brand"></label>
                         <select name="brand" id="brand_select" required>
                             <option value="">BRAND</option>
-                        {% for key_brand, brand in constant('App\\Model\\PartManager::BRAND') %}
+                        {% for key_brand, brand in constant('App\\Model\\PostManager::BRAND') %}
                             <option value="{{ key_brand + 1 }}">{{ brand }}</option>
                         {% endfor %}
                         </select>
@@ -56,7 +56,7 @@
                         <label for="location"></label>
                         <select name="location" id="location_select" data-rel="form-sample-phone" required>
                             <option value="">LOCATION</option>
-                        {% for location in constant('App\\Model\\PartManager::LOCATION') %}
+                        {% for location in constant('App\\Model\\PostManager::LOCATION') %}
                             <option value="{{ location }}">{{ location }}</option>
                         {% endfor %}
                         </select>
